Make jiant pip installable (#887)
* Create initial package for PyPI (pip) distribution

* Working PyPI package - can run tutorial experiment

This works for the tutorial experiment, but is still a work-in-progress.

* Add pip dependencies - matched environment.yml as closely as possible
* Move config into jiant package

* PEP 8 long strings changed to multi-line

* Update CI demo.conf path

* Use locally installed pip package for CircleCI

* Update call to ipdb

* Update config paths in non-python files

* Document release and distribution

* Name fix.

* Add note to 'Contributing'
