+++ jiant/config/defaults.conf
@@ -11,7 +11,7 @@
 //
 // ... or in a command line flag:
 //
-//   $ python main.py --config_file config/defaults.conf \
+//   $ python main.py --config_file jiant/config/defaults.conf \
 //         --overrides "exp_name = my_expt, run_name = run1, pretrain_tasks = \"mnli,qnli\""
 //
 // This file uses HOCON, which is a JSON/YAML-like format that supports
