+++ jiant/config/demo.conf
@@ -1,9 +1,9 @@
 // Fast demo config, used mainly for validating that an installation worked and that code
 // changes don't crash.
 // Run with:
-//   python main.py --config_file config/demo.conf
+//   python main.py --config_file jiant/config/demo.conf
 
-// This is not meant to yield good performance, but for reference, here are the results 
+// This is not meant to yield good performance, but for reference, here are the results
 // shown at the end of a typical run with jiant 1.0:
 
 // micro_avg: 0.285, macro_avg: 0.285, sts-b_corr: 0.285, sts-b_pearsonr: 0.297, sts-b_spearmanr: 0.273
