+++ jiant/config/edgeprobe/edgeprobe_cove.conf
