+++ jiant/config/edgeprobe/edgeprobe_existing.conf
@@ -1,8 +1,8 @@
 // Override config to eval an existing model. Intended to be included *after*
 // that model's params.conf. Usage:
 //   PARAM_FILE="/path/to/trained/model/params.conf"
-//   python main.py -c config/defaults.conf ${PARAM_FILE} \
-//     config/edgeprobe_existing.conf
+//   python main.py -c jiant/config/defaults.conf ${PARAM_FILE} \
+//     jiant/config/edgeprobe/edgeprobe_existing.conf
 //
 // See major_experiment_scripts/edgeprobe_example_run.sh
 // for usage example.
