+++ jiant/config/examples/copa_bert.conf
@@ -1,6 +1,6 @@
-// An example configuration for the CommitmentBank task with BERT.
+// An example configuration for the COPA task with BERT.
 // Run with:
-//   python main.py --config_file config/cb_bert.conf
+//   python main.py --config_file jiant/config/examples/copa_bert.conf
 
 // This imports the defaults, which can be overridden below.
 include "defaults.conf"
@@ -25,7 +25,7 @@ optimizer = bert_adam
 lr = 0.00001
 sent_enc = none
 sep_embs_for_skip = 1
-max_seq_len = 510
+max_seq_len = 512
 dropout = 0.1
 
 // Trainer setup for small tasks with BERT
