+++ jiant/config/examples/stilts_example.conf
@@ -2,7 +2,7 @@
 // for BERT -> MNLI -> RTE
 // For this example we use BERT-base.
 // Run with:
-//   python main.py --config_file config/examples/stilts_example.conf
+//   python main.py --config_file jiant/config/examples/stilts_example.conf
 
 include "../defaults.conf"
 pretrain_tasks = "mnli"
