+++ probing/README.md
@@ -21,12 +21,12 @@ mkdir -p $JIANT_DATA_DIR
 ```
 This should populate `$JIANT_DATA_DIR/edges` with directories for each task, each containing a number of `.json` files as well as `labels.txt`. For more details on the data format, see below.
 
-The main entry point for edge probing is [`jiant/main.py`](../main.py). The main arguments are a config file and any parameter overrides. The [`jiant/config/edgeprobe/`](../config/edgeprobe/) folder contains [HOCON](https://github.com/lightbend/config/blob/master/HOCON.md) files as a starting point for all the edge probing experiments.
+The main entry point for edge probing is [`jiant/main.py`](../main.py). The main arguments are a config file and any parameter overrides. The [`jiant/jiant/config/edgeprobe/`](../jiant/config/edgeprobe/) folder contains [HOCON](https://github.com/lightbend/config/blob/master/HOCON.md) files as a starting point for all the edge probing experiments.
 
 For a quick test run, use a small dataset like `spr2` and a small encoder like CoVe:
 ```sh
 cd ${PWD%/jiant*}/jiant
-python main.py --config_file config/edgeprobe/edgeprobe_cove.conf \
+python main.py --config_file jiant/config/edgeprobe/edgeprobe_cove.conf \
   -o "target_tasks=edges-spr2,exp_name=ep_cove_demo"
 ```
 This will keep the encoder fixed and train an edge probing classifier on the SPR2 dataset. It should run in about 4 minutes on a K80 GPU. It'll produce an output directory in `$JIANT_PROJECT_PREFIX/ep_cove_demo`. There's a lot of stuff in here, but the files of interest are:
@@ -54,9 +54,9 @@ You can use the `run/*_val.json` and `run/*_test.json` files to run scoring and 
   some of the features in `analysis.py`
 - [analyze_runs.py](analyze_runs.py) is a helper script to process a set of
   predictions into a condensed `.tsv` format. It computes confusion matricies for each label and along various stratifiers (like span distance) so you can easily and quickly perform further aggregation and compute metrics like accuracy, precision, recall, and F1. In particular, the `run`, `task`, `label`, `stratifier` (optional), and `stratum_key` (optional) columns serve as identifiers, and the confusion matrix is stored in four columns: `tp_count`, `fp_count`, `tn_count`, and `tp_count`. If you want to aggregate over a group of labels (like SRL core roles), just sum the `*_count` columns for that group before computing metrics.
-- [get_scalar_mix.py](get_scalar_mix.py) is a helper script to extract scalar 
+- [get_scalar_mix.py](get_scalar_mix.py) is a helper script to extract scalar
   mixing weights and export to `.tsv`.
-- [analysis_edgeprobe_standard.ipynb](analysis_edgeprobe_standard.ipynb) shows 
+- [analysis_edgeprobe_standard.ipynb](analysis_edgeprobe_standard.ipynb) shows
   some example analysis on the output of `analyze_runs.py` and `get_scalar_mix.py`. This mostly does shallow processing over the output, but the main idiosyncracies to know are: for `coref-ontonotes`, use the `1` label instead of `_micro_avg_`, and for `srl-ontonotes` we report a `_clean_micro_` metric which aggregates all the labels that don't start with `R-` or `C-`.
 
 ## Running the experiments from the paper
@@ -65,7 +65,7 @@ We provide a frozen branch, [`ep_frozen_20190723`](https://github.com/nyu-mll/ji
 
 Additionally, there's an older branch, [`edgeprobe_frozen_feb2019`](https://github.com/jsalt18-sentence-repl/jiant/tree/edgeprobe_frozen_feb2019), which is a snapshot of `jiant` as of the final version of the ICLR paper. However, this is much messier than above.
 
-The configs in `jiant/config/edgeprobe/edgeprobe_*.conf` are the starting point for the experiments in the paper, but are supplemented by a number of parameter overrides (the `-o` flag to `main.py`). We use a set of bash functions to keep track of these, which are maintained in [`jiant/scripts/edges/exp_fns.sh`](../scripts/edges/exp_fns.sh).
+The configs in `jiant/jiant/config/edgeprobe/edgeprobe_*.conf` are the starting point for the experiments in the paper, but are supplemented by a number of parameter overrides (the `-o` flag to `main.py`). We use a set of bash functions to keep track of these, which are maintained in [`jiant/scripts/edges/exp_fns.sh`](../scripts/edges/exp_fns.sh).
 
 To run a standard experiment, you can do something like:
 ```sh
