+++ scripts/demo.sh
@@ -2,4 +2,4 @@
 
 # Quick-start: use path_config.sh to set up your environment variables, then run this
 pushd "${PWD%jiant*}jiant"  # Make sure we're in the base jiant directory
-python main.py --config_file config/demo.conf
+python main.py --config_file jiant/config/demo.conf
