+++ scripts/nyu_cilvr_cluster.sbatch
@@ -8,7 +8,7 @@
 #SBATCH --constraint=gpu_12gb,pascal
 
 # Example usage:
-# JIANT_OVERRIDES="exp_name = main-multi-task-glue, pretrain_tasks = glue, run_name = noelmo-do2-sd1, elmo_chars_only = 1, dropout = 0.2" JIANT_CONF="config/defaults.conf" sbatch ../nyu_cilvr_cluster.sbatch
+# JIANT_OVERRIDES="exp_name = main-multi-task-glue, pretrain_tasks = glue, run_name = noelmo-do2-sd1, elmo_chars_only = 1, dropout = 0.2" JIANT_CONF="jiant/config/defaults.conf" sbatch ../nyu_cilvr_cluster.sbatch
 
 # Log what we're running and where.
 echo $SLURM_JOBID - `hostname` - $JIANT_OVERRIDES >> ~/jiant_machine_assignments.txt
